+++ CHANGELOG.md
@@ -43,6 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bug fixes
 
+- [#1150](https://github.com/alleslabs/celatone-frontend/pull/1150) Fix camel case on message
+
 ## v1.9.0
 
 ### Features